rename: 뒤로가기 버튼 공통 컴포넌트 폴더 생성후 이동
